feat: refactor api to be more dry
