+++ pages/api/discover/popular.js
@@ -1,16 +1,14 @@
-import fs from "fs";
-import path from "path";
-
 import connectMongo from "../../../config/mongo";
 import logger from "../../../config/logger";
 import Profile from "../../../models/Profile";
+import loadProfiles from "../../../services/profiles/loadProfiles";
 
 export default async function handler(req, res) {
   await connectMongo();
 
-  let popularProfiles = [];
+  let profiles = [];
   try {
-    popularProfiles = await Profile.find({
+    profiles = await Profile.find({
       username: { $nin: process.env.SHADOWBAN.split(",") },
     })
       .sort({ views: -1 })
@@ -19,31 +17,14 @@ export default async function handler(req, res) {
     logger.error(e, "failed to load profiles");
   }
 
-  if (popularProfiles.length === 0) {
+  if (profiles.length === 0) {
     return res.status(404).json([]);
   }
 
-  const directoryPath = path.join(process.cwd(), "data");
+  const profilesWithStats = await loadProfiles(
+    profiles.map((profile) => profile._doc)
+  );
 
-  const fullPopularProfiles = popularProfiles.flatMap((profile) => {
-    const filePath = path.join(directoryPath, `${profile.username}.json`);
-    try {
-      const user = JSON.parse(fs.readFileSync(filePath, "utf8"));
-
-      if (user.displayStatsPublic) {
-        return {
-          ...user,
-          ...profile._doc,
-        };
-      }
-
-      return [];
-    } catch (e) {
-      logger.error(e, `failed to get load profiles: ${filePath}`);
-      return [];
-    }
-  });
-
-  const selectedPopularProfiles = fullPopularProfiles.slice(0, 5);
+  const selectedPopularProfiles = profilesWithStats.slice(0, 5);
   res.status(200).json(selectedPopularProfiles);
 }
