+++ pages/api/discover/random.js
@@ -4,13 +4,14 @@ import path from "path";
 import connectMongo from "../../../config/mongo";
 import logger from "../../../config/logger";
 import Profile from "../../../models/Profile";
+import loadProfiles from "../../../services/profiles/loadProfiles";
 
 export default async function handler(req, res) {
   await connectMongo();
 
-  let randomProfiles = [];
+  let profiles = [];
   try {
-    randomProfiles = await Profile.aggregate([
+    profiles = await Profile.aggregate([
       { $sample: { size: 5 } },
       { $match: { username: { $nin: process.env.SHADOWBAN.split(",") } } },
     ]);
@@ -18,22 +19,11 @@ export default async function handler(req, res) {
     logger.error(e, "failed to load profiles");
   }
 
-  if (randomProfiles.length === 0) {
+  if (profiles.length === 0) {
     return res.status(404).json([]);
   }
 
-  const directoryPath = path.join(process.cwd(), "data");
-  const fullRandomProfiles = randomProfiles.flatMap((profile) => {
-    const filePath = path.join(directoryPath, `${profile.username}.json`);
-    try {
-      const user = JSON.parse(fs.readFileSync(filePath, "utf8"));
-
-      return { ...user, username: profile.username };
-    } catch (e) {
-      logger.error(e, `failed to get load profiles: ${filePath}`);
-      return [];
-    }
-  });
+  const fullRandomProfiles = await loadProfiles(profiles);
 
   res.status(200).json(fullRandomProfiles);
 }
