+++ pages/api/users/[username]/links/[url].js
@@ -1,12 +1,10 @@
-import fs from "fs";
-import path from "path";
-
 import connectMongo from "../../../../../config/mongo";
 import logger from "../../../../../config/logger";
 
 import Link from "../../../../../models/Link";
 import Profile from "../../../../../models/Profile";
 import Stats from "../../../../../models/Stats";
+import findOneByUsernameBasic from "../../../../../services/profiles/findOneByUsernameBasic";
 
 export default async function handler(req, res) {
   await connectMongo();
@@ -19,14 +17,11 @@ export default async function handler(req, res) {
       .json({ error: "Invalid request: GET request required" });
   }
 
-  // load profile json file and check link
-  const filePath = path.join(process.cwd(), "data", `${username}.json`);
-  let data = {};
-  try {
-    data = JSON.parse(fs.readFileSync(filePath, "utf8"));
-  } catch (e) {
-    logger.error(e, `failed loading profile username: ${username}`);
-    return res.status(404).json({ error: `ERROR ${username} not found` });
+  const data = findOneByUsernameBasic(username);
+
+  if (!data.username) {
+    logger.error(`failed loading profile username: ${username}`);
+    return res.status(404).json({ error: `${username} not found` });
   }
 
   if (
